+++ src/main/java/com/stevekung/moreplanets/planets/nibiru/client/renderer/tileentity/TileEntityNuclearWasteTankRenderer.java
@@ -31,7 +31,7 @@ public class TileEntityNuclearWasteTankRenderer extends TileEntitySpecialRendere
     @Override
     public void render(TileEntityNuclearWasteTank tile, double x, double y, double z, float partialTicks, int destroyStage, float alpha)
     {
-        Random rand = new Random(tile.getPos().getX() + tile.getPos().getY() * tile.getPos().getZ());
+        Random rand = new Random(tile.getPos().getX() + (long) tile.getPos().getY() * tile.getPos().getZ());
         float lightMapSaveX = OpenGlHelper.lastBrightnessX;
         float lightMapSaveY = OpenGlHelper.lastBrightnessY;
         GlStateManager.pushMatrix();
